cli/requirements.txt
requests==2.31.0 
infi.systray==0.1.12
win11toast==0.9
